Added SavedTagSelections Added ordering to chapter and chapter content Added seed data to enums and lookup tables

+++ TheCanalaveLibrary/TheCanalaveLibrary/Models/AcknowledgmentRole.cs
@@ -6,7 +6,7 @@ namespace TheCanalaveLibrary.Models;
 
 public partial class AcknowledgmentRole
 {
-    public byte AcknowledgmentRoleId { get; set; }
+    public short AcknowledgmentRoleId { get; set; }
 
     [Required]
     [MaxLength(256)]
